requirements.txt
fastapi
uvicorn
gradio
requests
python-docx
beautifulsoup4
PyMuPDF
faiss-cpu
sentence-transformers
